2024/src/test/kotlin/aoc/year2024/Day5Test.kt
package aoc.year2024

import aoc.library.solvePart1
import aoc.library.solvePart2
import io.kotest.matchers.ints.shouldBeExactly
import org.junit.jupiter.api.Test

class Day5Test {

  @Test
  fun part1TestInput() {
    val testInput = """
      47|53
      97|13
      97|61
      97|47
      75|29
      61|13
      75|53
      29|13
      97|29
      53|29
      61|53
      97|53
      61|29
      47|13
      75|47
      97|75
      47|61
      75|61
      47|29
      75|13
      53|13

      75,47,61,53,29
      97,61,53,29,13
      75,29,13
      75,97,47,61,53
      61,13,29
      97,13,75,29,47
    """.trimIndent()
    Day5.solvePart1(testInput) shouldBeExactly 143
  }

  @Test
  fun part1() {
    Day5.solvePart1() shouldBeExactly 5747
  }

  @Test
  fun part2TestInput() {
    val testInput = """
      47|53
      97|13
      97|61
      97|47
      75|29
      61|13
      75|53
      29|13
      97|29
      53|29
      61|53
      97|53
      61|29
      47|13
      75|47
      97|75
      47|61
      75|61
      47|29
      75|13
      53|13

      75,47,61,53,29
      97,61,53,29,13
      75,29,13
      75,97,47,61,53
      61,13,29
      97,13,75,29,47
    """.trimIndent()
    Day5.solvePart2(testInput) shouldBeExactly 123
  }

  @Test
  fun part2() {
    Day5.solvePart2() shouldBeExactly 5502
  }
}
